refactor fox / bacnet to scada package
