+++ zlib/conn.go
@@ -30,8 +30,8 @@ import (
 
 	"io"
 
-	"github.com/zmap/zgrab/ztools/bacnet"
 	"github.com/zmap/zgrab/ztools/ftp"
+	"github.com/zmap/zgrab/ztools/scada/bacnet"
 	"github.com/zmap/zgrab/ztools/ssh"
 	"github.com/zmap/zgrab/ztools/util"
 	"github.com/zmap/zgrab/ztools/x509"
