+++ ztools/scada/bacnet/common.go
