+++ ztools/scada/fox/fox.go
