+++ ztools/scada/fox/log.go
